lib: Add User interface module for the leds and the button
Note that the module has been modified.
This module is from the original firmware of the Thingy_52 like all
other files in the thingy_52 folder under libraries.
But it has beed modified so that all dependencies on BLE have been removed from the module as we are not using BLE in this application.